backend/requirements.txt
annotated-types==0.7.0
anthropic==0.83.0
anyio==4.12.1
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.6
distro==1.9.0
Django==6.0.3
django-cors-headers==4.3.1
django-filter==25.2
django-rest-passwordreset==1.5.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
docstring_parser==0.17.0
drf-yasg==1.21.15
groq==1.0.0
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jiter==0.13.0
packaging==26.0
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
pypdf==4.0.1
PyPDF2==3.0.1
python-decouple==3.8
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
requests==2.31.0
sniffio==1.3.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.12.0
